+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [#7272](https://github.com/rubocop-hq/rubocop/pull/7272): Show warning message if passed string to `Enabled`, `Safe`, `SafeAutocorrect`, and `AutoCorrect` keys in .rubocop.yml. ([@unasuke][])
 * [#7295](https://github.com/rubocop-hq/rubocop/pull/7295): Make it possible to set `StyleGuideBaseURL` per department. ([@koic][])
 * [#7301](https://github.com/rubocop-hq/rubocop/pull/7301): Add check for calls to `remote_byebug` to `Lint/Debugger` cop. ([@riley-klingler][])
+* [#7311](https://github.com/rubocop-hq/rubocop/pull/7311): Implement `Style/UnlessWithOperator` cop. ([@ChrisBr][])
 
 ### Bug fixes
 
@@ -4178,3 +4179,4 @@
 [@buehmann]: https://github.com/buehmann
 [@halfwhole]: https://github.com/halfwhole
 [@riley-klingler]: https://github.com/riley-klingler
+[@ChrisBr]: https://github.com/ChrisBr